suppost py2

+++ mrcnn/model.py
@@ -1,3 +1,4 @@
+from __future__ import division
 """
 Mask R-CNN
 The main Mask R-CNN model implementation.
@@ -2452,7 +2453,7 @@ class MaskRCNN():
         shift = np.array([wy1, wx1, wy1, wx1])
         wh = wy2 - wy1  # window height
         ww = wx2 - wx1  # window width
-        scale = np.array([wh, ww, wh, ww])
+        scale = np.array([wh, ww, wh, ww],dtype=np.float32)
         # Convert boxes to normalized coordinates on the window
         boxes = np.divide(boxes - shift, scale)
         # Convert boxes to pixel coordinates on the original image
